refactor: Use an explicit messagepack struct for the query
Messagepack is used in the input as well so we can be a bit more
explicit about it here as well.

This changes the query slightly as the crate used here manages to pack
`127` into a single byte instead of adding a leading `uint8` header but
the query is otherwise identical.